requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
speech_recognition>=3.10.0
google-generativeai>=0.3.0
deep-translator>=1.11.0
pdfminer.six>=20221105
python-docx>=0.8.11
python-dotenv>=1.0.0
spacy>=3.6.0
textblob>=0.17.1
openpyxl>=3.1.0
fpdf>=1.7.2
nest-asyncio>=1.5.6 